// src/api/model/addTestBody.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { AddTestBodySubTopicsItem } from './addTestBodySubTopicsItem';

export type AddTestBody = {
  date: string;
  subTopics: AddTestBodySubTopicsItem[];
  title?: string;
};