// src/api/model/subtopic.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */

export interface Subtopic {
  name: string;
  numberExercisesdoneByUser?: number;
  percentage?: number;
  subTopicId: number;
  totalExercises?: number;
}